lab-3-friday-team-three-edwin/rest/src/requirements.txt
pyramid==1.10
mysql-connector-python